PP-3174 Implement database migrations for ECS
Cater for 3 different startup scenarios
- dev migrate db and start
- start only
- migrate db only

Ensure Docker deploys the new script run-with-chamber script
Chamber executes docker-startup.sh which has logic for the
above-mentioned scenarios
We are keeping the docker-startup-with-db-migration.sh to
prevent breakages whilst we transition

TODO once we are happy with the new system
- remove the docker-startup-with-db-migration.sh and reference
in Dockerfile
- remove the RUN_APP VALUE variable in docker-startup.sh to ensure the
script fails in the event that the new variable is not set. The
RUN_MIGRATION will only be set by microservices that have a db.