BAU: Cleanup startup process
We don't use docker-startup-with-db-migration.sh anymore, so it
shouldn't exist, nor should it be in the Dockerfile.

We shouldn't have to pass in an environment variable to run the
application; it should run by default.

In order to run a migration and the app we set the environment
variables:

- RUN_MIGRATION_VALUE => true
- RUN_APP_VALUE       => true

In order to only run migrations we set:

- RUN_MIGRATION_VALUE => true
- RUN_APP_VALUE       => false (or anything that isn't true)

In order to only run the app don't set any environment variables.

I can't think of a case where we want to run neither the application nor
a migration.

solo @tlwr