PP-3714 Put facade over GoCardlessClientWrapper for testability
Some of the intermediate classes returned by the GoCardlessClient
are final and therefore not easy to mock. This makes it hard to
test the logic in the GoCardlessClientWrapper (which directly calls
the GoCardlessClient).

Work around this by putting a GoCardlessClientFacade between the
GoCardlessService and the GoCardlessClientWrapper. The
GoCardlessClientWrapper now does not do any processing or
conversion of the response objects it receives from the GoCardless
client. Instead, it just invokes the GoCardlessClient and returns
the result (a GoCardless-provided domain object) straight to the
GoCardlessClientFacade, which does the necessary processing and
conversion. The GoCardlessService now calls the
GoCardlessClientFacade rather than the GoCardlessClientWrapper.

Since the GoCardlessClientFacade talks only to the
GoCardlessClientWrapper (not the GoCardlessClient directly), it is
easy to mock the GoCardlessClientWrapper and make it return
appropriate GoCardless-provided domain objects (which are not
final) in tests. Meanwhile, the GoCardlessClientWrapper has very
little logic (just invokes the GoCardlessClient and returns the
result), making it a thin layer that is in less need of testing.