BAU: refactor like clause generation
using `LIKE_HOLDER.replace()` means that a reader of this code can't understand
what's happening without knowing the value of the constant. A nicer way to do this
whilst still keeping the code 'DRY' is to create a method to generate the like clause.

Using a method means that you can use the method name to convey the intent of the operation
and if you like, you can optimise the code in any way you want. The reader of the code can still
understand that a method `likeClause()` will generate a like clause, so they don't need
to worry about the implementation.