[WIP] Update pattern for remote + DB update
Basic sketch of what GC confirm looks like if we move to

1. Check transition is valid
2. @Transaction write placeholders (and event? Not done yet)
3. Create stuff with GC
4. @Transaction Update placeholders with GC data, write events and update status

Not done yet, and a bit messy, but don't want to go too much further without
some eyes. The main thing not done yet is "write events and update status", which
was originally being done at the beginning of the whole thing (as part of
TransactionConfirmService.confirm).

This pattern differs from what we were doing before which was roughly

1. Update status in DB
2. Create GC customer with GC and write to DB
3. Create GC Bank account with GC and write to DB
4. Create GC mandate with GC and write to DB
5. Optionally create GC payment with GC and write to DB

Th main advantage of new 'pattern': logically connected updates to DB
happen transactionally (crazy I know). If any call to GC fails, we will be left
with placeholder records, but no half completed nonsense.

Looking at this again, maybe the placeholder thing is not needed, just an event.

Hmmm.