PP-3942 Validate requests to validate bank account details
When processing a request to validate bank account details, first do
some basic validation of the request: ensure it contains sort_code
and account_number fields and that they are both numeric strings with
lengths in the correct ranges.

This prevents the code from blowing up unexpectedly if any of the
fields are missing and means we won’t try to perform a bank account
details lookup for obviously incorrect details.

This is consistent with the other endpoints, such as those for
mandate setup, which perform similar validation.

This change may alter behaviour because we will now return a non-200
response for some requests that we would previously refer to a lookup
service, which may have returned a 200 response indicating the
details are not valid, which we would have converted to one our 200
responses. However, this is an internal API only called by Direct
Debit frontend, which already only sends requests with sort_code and
account_number fields that are both numeric strings with lengths in
the correct ranges.