PP-3942 Introduce a SortCode object and use it in place of Strings
The SortCode object represents a UK bank account sort code with no
dashes or spaces e.g. "123456". It performs a check when it’s
constructed to ensure that it consists of exactly six Arabic
numerals. It’s safe to do this because SortCode objects are only
ever created to represent data in requests and our resources
ensure the sort codes match the correct format before the SortCode
objects are created.

Note that there’s no guarantee that a SortCode object actually
represents a real bank and branch.

The sortCode field in the Payer class remains a String rather than
becoming a SortCode object because we hash the sort code before
passing it to the Payer constructor (except in some tests), so it
isn’t actually a sort code. This also means that SortCode objects
are never persisted or depersisted: they are entirely ephemeral.