PP-3942 Introduce AccountNumber and use it in place of Strings
The AccountNumber object represents a UK bank account number with no
dashes or spaces e.g. "12345678". It performs a check when it’s
constructed to ensure that it consists of six to eight Arabic
numerals inclusive. It’s safe to do this because AccountNumber
objects are only ever created to represent data in requests and our
resources ensure the account numbers match the correct format before
the AccountNumber objects are created.

Note that there’s no guarantee that an AccountNumber object actually
represents a real bank account.

The accountNumber field in the Payer class remains a String rather
than becoming an AccountNumber object because we hash the account
number before passing it to the Payer constructor (except in some
tests), so it isn’t actually an account number. This also means that
AccountNumber objects are never persisted or depersisted: they are
entirely ephemeral.