BAU: Make tests which use DB work with docker for mac
Docker for mac is connected to over a unix socket rather than TCP. The postgres setup code
attempts to calculate the hostname of the postgres server. Rather than doing that, we can use
the docker class itself, which knows the hostname that it's running on. This means we can delete
a bunch of code.