PP-3962 Read in gateway account access token and organisation
Add accessToken and organisation fields to the GatewayAccount
object and populate them with the values of the (nullable)
access_token and organisation columns when de-persisting a
GatewayAccount entity from the database.

Introduce specific types to represent access tokens and
organisation identifiers rather than using strings.

No changes to GatewayAccountParser because we do not currently
support specifying an access token or organisation identifier when
creating a gateway account from a JSON payload.

No changes to GatewayAccountResponse because we there’s no need for
anyone outside of Direct Debit connector to know about these values
(in particular, the access token must be kept secret).