PP-3965 Disable manadatory organisation
- At the moment we have a sandbox account with GoCardless and they
do not send the field `organisation` in the request so this makes our
webhooks to fail. Following stories will upgrade our GoCardless account
to a Partner account and then we will have `organisation` in the payload.
To unblock other stories I left the logic in place, but replaced the
field with `null`, in case it doesn't exist.
- This should be enforced in later stories.