PP-4043 Link GoCardless account to GOV.UK Pay account
- Added a resource to create, disable, get oauth tokens
- Added service for GoCardless partner app token
- Added Dao for GoCardless partner app token
- Added mocked resource test, that will assert the response code, headers
and the entity. I believe that spinning up a new HtpServlet environment and
a db are an overkill given that there are tests: the service in isolation,
the Dao with the database, the resource should be tested against Pact as
a producer.
- Used a custom Dropwizard NonEmptyStringParam type, which treats empty
strings as missing. Also is easy to test a resource in isolation.