PP-4660: Remove PayLoggerFactory
We are already using MDC to embed the request ID into the logging pipeline.
It appears in our logs via the %X placeholder in our logformat string so
don't need to create proxies to the SLF4J LoggerFactory in order to intercept
logger calls.

Remove this extra layer of indirection and duplicate logging of request IDs.