Correct tests to enable DropWizard upgrade
- Correct `DatabaseTestHelper.getTokenByMandateExternalId` and `getTrasnactionsForMandate`
to return only contents of `transactions` table. New version of JDBI appears to use
relfection which complains if two cols in the same result set have the same name
which makes sense since the returned map cannot know which of the multiple keys to
have used. This occurs on multiple fields if joining transactions and mandates tables.

- Correct the assertion to test the status of the transactions which should be
PENDING and not the mandate state which is SUBMITTED. This error became apparent when
the `getTransactionsForMandate` was changed to only return transaction details and
not include mandate detail (both have `state` cols and therefor the returned Map
is incorrect and the test was wrong).