PP-5232 Remove reliance on exception to control logic
- Make DirectDebitStatesGraph return Optional<State> rather than throwing an
exception if no next state can be found. This exception was being relied upon
to control logic with confusing methods such as `void canUpdateStateFor` which
appears to be asking a question but with a return type of `void` was relying upon
the exception.
- Move the throwing of exception higher in the call stack and in the context of
the check that was being performed. e.g canUpdateState ? do something : throw exception.