PP-5231 Introduce ServiceMandateReference
ServiceMandateReference represents the reference assigned to a
mandate by a government service (set when the service sends the
request to us to create the mandate). It wraps a string.

This isn’t quite as nice as it could be because
CreateMandateRequest (create an on-demand mandate) and
CreatePaymentRequest (either create a one-off mandate and payment
to be collected from it, or collect a payment from an on-demand
mandate) both implement CreateRequest and CreatePaymentRequest
also implements CollectRequest, which both define the method
getReference() to return a String object (it represents either a
service’s reference for a mandate or a service’s reference for a
payment or both at the same time).