PP-5231 Introduce MandateBankStatementReference
MandateBankStatementReference represents the reference for a
mandate that may be shown on the paying user’s bank statement. It
wraps a string.

Having this type seems particularly valuable because it tends to
be called either mandateReference or goCardlessReference and it’s
not particularly obvious these represent the same thing when
they’re both plain strings. As a result, introducing this was
satisfying and changing the types of variables from strings to this
new class until everything compiled was particularly satisfying.
Like when you’re cutting wrapping paper and the scissors start to
glide and you don’t have to make the snipping action any more.

A few classes that represent JSON responses assumed that fields
that now have the type MandateBankStatementReference could be
null (in equals(…) methods etc.): as far as I can tell, this isn’t
the case so the they’ve been updated to assume the fields will
never be null.