PP-5234: Make service name nullable
The service name was removed in the call to dd-connector at
https://github.com/alphagov/pay-selfservice/pull/1482.

Before I remove the service name from GatewayAccountResponse and
CreateGatewayAccountRequest the service_name column in the gateway_accounts
table needs to be made nullable.

Once I remove the service name from GatewayAccountResponse and
CreateGatewayAccountRequest in another PR I'll be delete to drop the column
completely.