PP-5298 Include details for payment state in responses
Include a "details" field in the state object for Payment responses.
This will contain the "details[cause]" obtained from a Go Cardless event.

Currently this is always set to a hardcoded value and needs to be wired
up to an actual value retrieved from the database when event processing
is implemented properly.

Co-authored-by: Rory Malcolm <rory.malcolm@digital.cabinet-office.gov.uk>