PP-5428 Separate out payment creation and collection
Untangle creating a payment from collecting a payment in
PaymentService by introducing a new CollectService to orchestrate
the overall operation. This makes the code easier to test and
reason about.

Remove the look-up of the gateway account from the URL because it
was never actually used (though it would throw an exception if the
gateway account didn’t exist — it no longer does, so this is a
behavior change).

Make it so that when the mandate is looked up, it’s only returned
if it matches the gateway account (this change should be applied
to several other places too).