PP-5318 Use WrappedStringValue where appropriate: code reuse FTW
Make specific types that wrap strings extend WrappedStringValue so
we don’t have identical implementations of equals(…), hashCode()
etc.

This change also makes PaymentProviderMandateId and
PaymentProviderSandboxId stricter when it comes to equality:
SandboxMandateId.valueOf("foo") and
GoCardlessMandateId.valueOf("foo") will no longer be considered
to be equal (which is what we want because they’re not). Fix a
test that relied on the buggy behaviour.

This may change the hash codes of some objects but since they’re
not shared between application nodes or persisted across
application restarts, this is fine.