PP-5371 Rationalise organisation identifier in GoCardlessEvent
GoCardlessEvent had two fields for the organisation identifier
("links"."organisation" in the JSON): organisationIdentifier and
linksOrganisation. The former was a GoCardlessOrganisationId while
the latter was a String. While we parsed both out of the incoming
webhook event JSON, only linksOrganisation got persisted to and
read from the database. In other places, we were inconsistent
about which one we used.

Standardise on linksOrganisation (for consistency with similar
fields in GoCardlessEvent) but make the type a
GoCardlessOrganisationId rather than a String.

Remove code that tries to account for events that don’t have an
organisation ID: all incoming events should do have an organisation
ID (because they’re from an accounts linked to our partner app):
there might still be some in the database but since we’re not live
yet we can always clear them all out.