PP-5302 Modify PaymentViewReponse to be generic for re-use.
- Rename `PaymentViewResponse` to `SearchResponse` and make it generic to permit
paramaterisation of the contents of the `List` of results. This means it can be
reused for Mandates as well as Payments.
- Move into the common package.
- Refactor `PaymentSearchService` for clarity
- Rename `ViewPaginationBuilder` to `LinksForSearchResult` to hopefully better
explain what it does, also remove the notion of it being a builder because it isn't
and make as many fields `final` as possible.
- Note that within existing logic for `LinksForSearchResult.buildLinks()` it mutates
the page number within the `searchParams` it was provided to create the links
which is a little confusing. In future PR we hope to change this. This is why
`long originalPageNumber = validatedSearchParams.getPage();` has been added for now
within `PaymentSearchService.getPaymentSearchResponse()`.