PP-5318 MandateLookupKey for updating mandate state from GC events
GoCardless mandate IDs aren’t unique, so to retrieve a mandate from
the database, we have use the mandate ID in combination with the
GoCardless organisation ID. Sandbox mandate IDs don’t have this
limitation and sandbox accounts don’t have organisation IDs.

Make things simpler for ourselves by introducing a
MandateUpdateService, which acts as a façade over the MandateDao
and takes a MandateLookupKey object. Both
GoCardlessMandateIdAndOrganisationId (a value type that wraps a
GoCardless mandate ID and organisation ID) and SandboxMandateId
implement the MandateLookupKey marker interface. The
MandateUpdateService uses some fantastically non-polymorphic code
to determine the precise runtime type of the MandateLookupKey it’s
been passed and calls the appropriate method on the MandateDao
depending on whether the lookup needs to take account of the
organisation ID or not.

Make use of the new MandateUpdateService when updating the status
of a mandate in response to a GoCardless webhook. This means we
can pass it a GoCardlessMandateIdAndOrganisationId (rather than a
GoCardlessMandateId and a GoCardlessOrganisationId).

Spread more widely, this pattern will allow us to remove
PaymentProviderServiceId, which is the interface implemented by
GoCardlessOrganisationId in order to maintain the fiction that
all payment providers have this concept of organisation IDs.