PP-5302 Split the search method from MandateQueryService into new MandateSearchService
In a bid to have smaller services this moves the mandate search funtion into a
new MandateSearchService. This also helps keep associated test files more targeted
and easier to understand.

Renamed MandateSearchResourceIT to MandateResourceSearchIT to help people find it
from MandateResourceIT, e.g. by searching on the class name that the test is
associated with.