PP-5431 Payment status updater that reads GoCardless webhook events
After receiving a GoCardless webhook event, update the state of any
affected payments by looking at the latest stored event (not just
the one that was just received) and calculating the state from the
event’s action. This doesn’t enforce any idea of valid state
transitions.

Bypass code that caused us to set all payments as failed if we
received an event with a mandate failed action; we should get a
separate payment event from GoCardless so this is unnecessary.