PP-5432 Read all webhook events when updating sandbox payment state
After processing a sandbox webhook, update the states of the
affected payments (which is really all of them) by examining all
stored events rather than the one we just received. No state
transition restrictions are enforced.

This brings sandbox into line with GoCardless in terms of its
(theoretical) ability to handle out-of-order webhooks.