PP-5433 Lookup payment to calculate state
We need to take into account GovUkPayEvents when calculating the state
of a payment. We look up GovUkPayEvents by the payment id
rather than the payment provider identifiers as a payment may not have
the payment provider details set. To accomplish this we need to have the
Payment in the calculator rather than just the payment provider
identifiers.