PP-5433 Include GovUkPayEvents in Sandbox payment state calculation
When calculating the state of a Sandbox payment, find the most
recent Sandbox and GovUkPayEvent then use whichever has a later
timestamp to calculate the state.

Add ...EventToPaymentStateMapper classes to share the GovUkPayEvent
mapping between the GoCardless and Sandbox calculators. Add classes for
GoCardless and Sandbox for consistency, as this was done for the mandate
state calculation.