PP-5478 Unhardcode webhook signatures in webhook integration tests
In integration tests that hit the GoCardless webhook endpoint but
aren’t trying to test the signature verification code, calculate
what the signature should be transparently in the test, rather
than hardcoding the expected value. This makes the tests less
brittle because there’s no hardcoded signature to update whenever
the test webhook changes.

Separate out the code that calculates a webhook signature from the
code that verifies a signature against the expected value in order
to facilitate this.