PP-5455 Add REFUNDS and SUBSCRIPTIONS to GoCardlessResourceType
Add REFUNDS and SUBSCRIPTIONS to GoCardlessResourceType enum. When
we receive a webhook event from GoCardless with a "resource_type"
of "refunds" or "subscriptions", we will convert it to a value
from the enum before persisting to the database. Without REFUNDS
and SUBSCRIPTIONS in the enum, we were persisting the default
UNHANDLED (which has now been removed).