PP-5483 Remove ApacheConnector from our JerseryClient set-up
It appears that ApacheConnector does not respect the http.nonProxyHosts setting
which is causing problems when DDConnector attempts to call Adminusers (receives
a 403 forbidden from the proxy which it shouldn't be attempting to go via). This
switches our client implementation to the same as PublicApi. There is no setting
in DDConnector config to control whether or not to use SSL (as there is in PublicApi)
and we don't see a need to turn this on or off so it will always set-up SSL.

Change the timeout from 9000ms to 5000ms as 9000ms seems too long but since we
do not have any retry with gocardless it should have a comfortable margin, as
discussed with Alex.

with @RoryMalcolm