PP-5493 Check if mandate has provider ID before creating payment
Check if a mandate has a provider ID before creating the payment
and inserting it into the database. This means that if for some
reason we get a request to collect a payment from a mandate that
does not have a provider ID (hasn’t been submitted to the provider
yet), we return an error earlier and won’t have a useless payment
sitting around in the database.