When processing a NotFoundException, log at WARN rather than ERROR
Many NotFoundExceptions are caused by input that’s outside of our
control, so log at WARN (which doesn’t create a support ticket)
rather than ERROR (which does).

When we have time, we might want to make any NotFoundExceptions
that are not caused by external input log errors again, as this
is more indicative of a problem we can do something about.