Tweak logging for errors validating bank details with GoCardless
When there’s a GoCardlessException while validating bank account
details, don’t log two lines, one with the GoCardlessException
message immediately, and one without the message when the resultant
InternalServerErrorException. Instead do all the logging when
mapping the InternalServerErrorException and enhance this so it
always includes the names of and messages from the wrapped
exceptions.

We can replace this with something even better when we have
structured logging in place.