PP-5735: Update json logging configuration
We can remove the timeZone = UTC bit as it's UTC by default. The timeZone
corresponds to the timeZone field in the
`io.dropwizard.lgging.AbstractAppenderFactory` but this is never actually used
when building the layout of the LogstashConsoleAppenderFactory.

We can remove the timestamp format as the LogstashConsoleAppenderFactory in
pay-java-commons logs the ISO-8601 formatted timestamp by default.

We can also remove the

```
layout:
  type: json
```
config as the LogstashConsoleAppenderFactory logs in json by default.

Example log produced by this configuration:
```
{
  "@timestamp": "2019-10-16T11:07:13.070+01:00",
  "@version": "1",
  "message": "Registering jersey handler with root path prefix: /",
  "logger_name": "io.dropwizard.server.DefaultServerFactory",
  "thread_name": "main",
  "level": "INFO",
  "level_value": 20000,
  "label": "directdebit-connector"
}
```

From local testing I've also verified MDC values are logged and structured
argument values in log lines are logged as a key value pair in the json log.
For example,
```
log.info("This is a log line", StructuredArguments.kv("gateway_account_id","abc"))
```
will produce
```
{
  "@timestamp": "2019-10-16T11:07:13.070+01:00",
  "@version": "1",
  "message": "This is a log line",
  "gateway_account_id": "abc"
  "logger_name": "io.dropwizard.server.DefaultServerFactory",
  "thread_name": "main",
  "level": "INFO",
  "level_value": 20000,
  "label": "directdebit-connector"
}
```